Refactor cobuild into a standalone library (#47)
* refactor cobuild into a standalone library
* fix simulator bug on witness
* cleanup comments
* update clang-format docker
